Update javadoc, clean up test mocking
